requirements.txt
flask
flask-cors
matplotlib
pandas
python-dotenv
requests
scikit-learn
